Using a simple custom user model.
At this moment in time, there is no need to create a fully custom user
model. Instead, I am inheriting from AbstractUser as that already has decent
defaults. We are still using a custom user model so in the future I can easily
customize the user model if need be.